meetingIQ_backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
distro==1.9.0
Django==4.2.23
django-cors-headers==4.7.0
djangorestframework==3.16.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.95.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
regex==2024.11.6
requests==2.32.4
sniffio==1.3.1
sqlparse==0.5.3
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
